requirements.txt
black==22.12.0
cachetools==5.2.0
click==8.1.3
google-auth==2.14.1
google-auth-oauthlib==0.7.1
gspread==5.7.1
oauthlib==3.2.2
pathspec==0.10.3
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests-oauthlib==1.3.1
rsa==4.9
